refactor(search): Extract searchQueryFromUrl and useUploadedTraces
- Move URL→SearchQuery parsing into searchQueryFromUrl() in url.ts,
  co-located with getUrlState/getUrl/isSameQuery
- Extract uploaded-trace cache management into useUploadedTraces() hook
  in SearchTracePage/ (search-specific, not in shared hooks/)
- SearchTracePage/index.tsx is now focused on composition and rendering
